+++ bee-hive/bee_hive/__init__.py
@@ -1,6 +1,3 @@
-
-__all__ = [
-    "Agent",
-    "Workflow",
-    "Interface",
-]
+# SPDX-License-Identifier: Apache-2.0
+from .workflow import Workflow
+from .bee_agent import BeeAgent
