+++ bee-hive/bee_hive/agent_factory.py
@@ -1,8 +1,8 @@
-
+# SPDX-License-Identifier: Apache-2.0
 from enum import Enum
 from typing import Callable, Type, Union
-from bee_hive.bee_agent import BeeAgent
-from bee_hive.crewai_agent import CrewAIAgent
+from .bee_agent import BeeAgent
+from .crewai_agent import CrewAIAgent
 
 class AgentFramework(Enum):
     """Enumeration of supported frameworks"""
