+++ bee-hive/bee_hive/workflow.py
@@ -2,15 +2,15 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import os
-from bee_hive.step import Step
-from bee_hive.agent_factory import AgentFramework
 import dotenv
+from .step import Step
+from .agent_factory import AgentFramework
 
 # TODO: Refactor later to factory or similar
-from bee_hive.crewai_agent import CrewAIAgent
-from bee_hive.bee_agent import BeeAgent
-from bee_hive.mock_agent import MockAgent
-from bee_hive.agent import restore_agent
+from .crewai_agent import CrewAIAgent
+from .bee_agent import BeeAgent
+from .mock_agent import MockAgent
+from .agent import restore_agent
 
 dotenv.load_dotenv()
