+++ bee-hive/tests/bee/test_bee.py
@@ -6,8 +6,9 @@ import yaml
 import os
 from pytest_mock import mocker
 
-from bee_hive.workflow import Workflow
-from bee_hive.bee_agent import BeeAgent
+# TODO consider moving to same directory as BeeAgent source
+from bee_hive import Workflow
+from bee_hive import BeeAgent
 
 dotenv.load_dotenv()
