+++ bee-hive/tests/crewai/test_crewai.py
@@ -2,10 +2,10 @@
 
 from unittest import TestCase
 import dotenv
-import yaml
 import os
+import yaml
 
-from bee_hive.workflow import Workflow
+from bee_hive import Workflow
 
 dotenv.load_dotenv()
